pyproject.toml
[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core>=1.0.0"]

[tool.poetry]
authors = ["Raphael Krupinski <rafalkrupinski@users.noreply.github.com>"]
description = "Python async OpenAPI client library generator"
license = "AGPL-3.0"
name = "lapidary-render"
version = "0.8.0"
packages = [{ include = "lapidary", from = "src" }]
readme = 'Readme.md'
repository = 'https://github.com/python-lapidary/lapidary'
include = [{ path = 'src/lapidary/render/templates' }]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Operating System :: OS Independent",
    "Topic :: Software Development :: Code Generators",
    "Typing :: Typed",
]

[tool.poetry.dependencies]
python = "^3.9"
black = "^23.10.1"
inflection = "^0.5.1"
Jinja2 = "^3.1.2"
jsonpatch = "^1.32"
lapidary = {path = "../lapidary", develop = true}
pydantic = "^2.5.2"
python-mimeparse = "^1.6.0"
PyYAML = "^6.0"
rybak = {path = "../../rybak", extras = ["jinja"], develop = true}
tomlkit = "^0.11.4"
typer = "^0.6.1"

[tool.poetry.scripts]
lapidary = "lapidary.render:app"

[tool.poetry.group.dev.dependencies]
pytest = "^7.1.3"

[tool.ruff]
target-version = "py312"
extend-select = ['I', 'UP']
extend-exclude = ['src/lapidary/render/templates/']
extend-ignore = ['F401', 'F403', 'F405']
line-length = 120

[tool.ruff.format]
quote-style = "single"

[tool.mypy]
mypy_path = "src"
namespace_packages = true
python_version = "3.12"
exclude = ["src/lapidary/render/templates"]

[tool.refurb]
python_version = "3.12"
ignore = ["115", "126"]
